Implementation for OkHttpClient
