+++ android/core/src/main/java/com/stadiamaps/ferrostar/core/Extensions.kt
@@ -11,23 +11,6 @@ import uniffi.ferrostar.getRoutePolyline
 
 @Throws fun Route.getPolyline(precision: UInt): String = getRoutePolyline(this, precision)
 
-@Throws
-fun RouteRequest.toOkhttp3Request(): Request {
-  val headers: Map<String, String>
-  return when (this) {
-        is RouteRequest.HttpPost -> {
-          headers = this.headers
-          Request.Builder().url(url).post(body.toRequestBody())
-        }
-
-        is RouteRequest.HttpGet -> {
-          headers = this.headers
-          Request.Builder().url(url).get()
-        }
-      }
-      .apply { headers.map { (name, value) -> header(name, value) } }
-      .build()
-}
 
 /** A neutral bounding box type, which is not dependent on any particular map library. */
 data class BoundingBox(
