+++ android/core/src/main/java/com/stadiamaps/ferrostar/core/http/HttpClient.kt
@@ -1,5 +1,8 @@
 package com.stadiamaps.ferrostar.core.http
 
+import okhttp3.OkHttpClient
+import okhttp3.Request
+import okhttp3.RequestBody.Companion.toRequestBody
 import uniffi.ferrostar.RouteRequest
 
 
@@ -17,3 +20,47 @@ interface IHttpClient {
     suspend fun call(request: RouteRequest): IResponse
 }
 
+
+class FerrostarOkHttpClient(private val client: OkHttpClient) : IHttpClient {
+    override suspend fun call(request: RouteRequest): IResponse {
+        val okHttpRequest = request.toOkhttp3Request()
+
+        val res = client.newCall(okHttpRequest).await()
+
+        return res.let { response ->
+            object : IResponse {
+                override val isSuccessful: Boolean
+                    get() = response.isSuccessful
+
+                override val code: Int
+                    get() = response.code
+
+                override fun bodyBytes(): ByteArray? = response.body?.bytes()
+            }
+        }
+    }
+
+    companion object {
+        fun OkHttpClient.toFerrostarOkHttpClient(): FerrostarOkHttpClient {
+            return FerrostarOkHttpClient(this)
+        }
+
+        @Throws
+        fun RouteRequest.toOkhttp3Request(): Request {
+            val headers: Map<String, String>
+            return when (this) {
+                is RouteRequest.HttpPost -> {
+                    headers = this.headers
+                    Request.Builder().url(url).post(body.toRequestBody())
+                }
+
+                is RouteRequest.HttpGet -> {
+                    headers = this.headers
+                    Request.Builder().url(url).get()
+                }
+            }
+                .apply { headers.map { (name, value) -> header(name, value) } }
+                .build()
+        }
+    }
+}
